Refactor main script (again)
