+++ index.html
@@ -8,10 +8,7 @@
     <meta name="color-scheme" content="dark light" />
     <title>Nathan Doolan</title>
 
-    <script type="module" src="/src/styles.ts"></script>
-    <script type="module" src="/src/analytics.ts"></script>
-    <script type="module" src="/src/projects.ts"></script>
-    <script type="module" src="/src/copyright.ts"></script>
+    <script type="module" src="/src/index.ts"></script>
     <script type="module" src="/src/theme.ts"></script>
     <script type="module" src="/src/logo-link-web-component.ts"></script>
 </head>
