--- src/analytics.ts
@@ -1,3 +1,0 @@
-import { inject } from "@vercel/analytics";
-
-inject();
