+++ src/copyright.ts
@@ -1,8 +1,3 @@
 export const replaceWithCurrentYear = (input: string, match: string): string => {
     return input.replace(match, new Date().getFullYear().toString());
 };
-
-const copyright = document.getElementById("copyright");
-if (copyright) {
-    copyright.innerHTML = replaceWithCurrentYear(copyright.innerHTML, "{current year}");
-}
