+++ src/projects.ts
@@ -301,65 +301,3 @@ export const projectIntoTemplate = (
 
     return templateClone;
 };
-
-export const appendRandom = async (parent: HTMLElement, ...elements: DocumentFragment[]) => {
-    for (const element of elements) {
-        const children = parent.children;
-        const randomIndex = Math.floor(Math.random() * (children.length + 1));
-
-        if (randomIndex === children.length) {
-            // If the random index is equal to the length, append at the end
-            parent.append(element);
-        } else {
-            // Otherwise, insert at the randomly determined index
-            parent.insertBefore(element, children[randomIndex]);
-        }
-    }
-};
-
-const loadProjects = async () => {
-    // Create project-gallery loader
-    const loader = document.createElement("span");
-    loader.classList.add("loader");
-    document.getElementById("projects")?.append(loader);
-
-    // Load items into gallery
-    const gallery = document.getElementById("project-gallery");
-    const template = document.getElementById("project-template") as HTMLTemplateElement | undefined;
-
-    if (gallery && template) {
-        const githubPage = await fetchData("/proxy/github");
-        const githubProjects = scrapeGithub(githubPage).map((p) =>
-            projectIntoTemplate(p, template),
-        );
-        await appendRandom(gallery, ...githubProjects);
-
-        const bggPage = await fetchData("/proxy/boardgamegeek");
-        const bggRawProjects = scrapeBgg(bggPage);
-
-        // Get higher resolution image from bgg xmlapi
-        for (const project of bggRawProjects) {
-            const id = project.url
-                ?.toString()
-                ?.split("/")
-                .find((v) => v.match(/\d+/g));
-            const gameXml = await fetchData(`/xmlapi/boardgamegeek/${id}`, "text/xml");
-            upgradeBggImage(project, gameXml);
-        }
-
-        const bggProjects = bggRawProjects.map((p) => projectIntoTemplate(p, template));
-
-        await appendRandom(gallery, ...bggProjects);
-
-        const cults3dPage = await fetchData("/proxy/cults3d");
-        const cults3dProjects = scrapeCults3d(cults3dPage).map((p) =>
-            projectIntoTemplate(p, template),
-        );
-        await appendRandom(gallery, ...cults3dProjects);
-    }
-
-    // remove loader
-    loader.remove();
-};
-
-loadProjects();
