--- src/styles.ts
@@ -1,1 +1,0 @@
-import "./styles.css";
